New config option: "allow_empty_feeds" (bug 569)
If this option is set to True, gPodder will accept feeds
that do not have any episodes listed. This is useful for
cases where podcasts do not have episodes throughout the
year, but will have new episodes eventually.

Thanks to Steve for reporting this bug in Bugzilla.

+++ src/gpodder/model.py
@@ -132,7 +132,7 @@ class PodcastChannel(PodcastModelObject):
 
     @classmethod
     def load(cls, db, url, create=True, authentication_tokens=None,\
-            max_episodes=0, download_dir=None):
+            max_episodes=0, download_dir=None, allow_empty_feeds=False):
         if isinstance(url, unicode):
             url = url.encode('utf-8')
 
@@ -149,8 +149,9 @@ class PodcastChannel(PodcastModelObject):
             tmp.update(max_episodes)
             tmp.save()
             db.force_last_new(tmp)
-            # Subscribing to empty feeds should yield an error
-            if sum(tmp.get_statistics()) == 0:
+            # Subscribing to empty feeds should yield an error (except if
+            # the user specifically allows empty feeds in the config UI)
+            if sum(tmp.get_statistics()) == 0 and not allow_empty_feeds:
                 tmp.delete()
                 raise Exception(_('No downloadable episodes in feed'))
             return tmp
